Fix Entity attributes

+++ be/src/main/java/com/sba301/code/be/repository/AccountRepository.java
@@ -1,9 +1,9 @@
 package com.sba301.code.be.repository;
 
-import com.sba301.code.be.entity.Account;
+import com.sba301.code.be.model.entity.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Integer> {
+public interface AccountRepository extends JpaRepository<Account, Long> {
 }
