+++ be/src/main/java/com/sba301/code/be/repository/CategoryRepository.java
@@ -1,9 +1,9 @@
 package com.sba301.code.be.repository;
 
-import com.sba301.code.be.entity.Category;
+import com.sba301.code.be.model.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 }
