+++ be/src/main/java/com/sba301/code/be/repository/OrderDetailRepository.java
@@ -1,9 +1,9 @@
 package com.sba301.code.be.repository;
 
-import com.sba301.code.be.entity.OrderDetail;
+import com.sba301.code.be.model.entity.OrderDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderDetailRepository extends JpaRepository<OrderDetail, Integer> {
+public interface OrderDetailRepository extends JpaRepository<OrderDetail, Long> {
 }
