+++ be/src/main/java/com/sba301/code/be/repository/OrderRepository.java
@@ -1,9 +1,9 @@
 package com.sba301.code.be.repository;
 
-import com.sba301.code.be.entity.Order;
+import com.sba301.code.be.model.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
