+++ be/src/main/java/com/sba301/code/be/repository/RoleRepository.java
@@ -1,9 +1,9 @@
 package com.sba301.code.be.repository;
 
-import com.sba301.code.be.entity.Role;
+import com.sba301.code.be.model.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
 }
