+++ be/src/main/java/com/sba301/code/be/service/AccountService.java
@@ -1,6 +1,6 @@
 package com.sba301.code.be.service;
 
-import com.sba301.code.be.entity.Account;
+import com.sba301.code.be.model.entity.Account;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -8,8 +8,8 @@ import java.util.List;
 @Service
 public interface AccountService {
     public List<Account> getAllAccounts();
-    public Account getAccountById(int accountId);
+    public Account getAccountById(Long accountId);
     public Account createAccount(Account account);
-    public Account updateAccount(int accountId, Account account);
-    public void deleteAccount(int accountId);
+    public Account updateAccount(Long accountId, Account account);
+    public void deleteAccount(Long accountId);
 }
